python-agent/requirements.txt
# ServWatch Python Agent Requirements

# WebSocket client
python-socketio>=5.0.0
requests>=2.28.0

# System monitoring
psutil>=5.9.0

# NVIDIA GPU monitoring
nvidia-ml-py>=12.0.0

# Optional: Advanced monitoring
# py3nvml>=0.2.7
# GPUtil>=1.4.0
# pynvml>=0.1.0
